feat(migrations): :sparkles: add migration for orders table
This commit adds a migration file `2024_08_26_200728_create_orders_table.php` under the `database/migrations` directory. The migration creates the `orders` table with columns for `total_price`, `status`, `session_id`, and foreign keys for `UserAddress`, `created_by`, and `updated_by`. This migration is necessary to support the functionality related to orders in the application.