feat(controllers): :sparkles: add store method to ProductController
This commit adds a new method called `store` to the `ProductController` in the `app/Http/Controllers/Admin/ProductController.php` file. The `store` method is responsible for creating a new product based on the user input and saving it to the database. It also handles the uploading of product images, if any are provided.

The `store` method receives a `Request` object as a parameter and uses it to retrieve the necessary data for creating a new product. It sets the product's attributes such as title, price, quantity, description, category_id, and brand_id based on the user input.

If the user has uploaded product images, the method saves them to the database and moves the uploaded files to the `product_images` directory. Each image is associated with the newly created product.

After successfully creating the product, the method redirects the user to the `admin.products.index` route and displays a success flash message.

This commit enhances the functionality of the `ProductController` and improves the user experience when creating new products.