feat(controllers): :sparkles: implement index method in CartController.php
This change implements the `index` method of the `CartController`. It handles viewing of the shopping cart. If user is logged in; it fetches `cartItems` and `userAddresses`  from model and if `cartItems` is greater than 0, passes the the fetched data to the `CartList` component. Otherwise, if the user is a guest user it fetches the `cartItems` from the cookie. If the `cartItems` is greater then 0, passes the `cartItems` to the `CartList` component or if the `cartItems` is empty it redirects back.